+++ apis/spec/v1alpha3/groupversion_info.go
@@ -28,14 +28,6 @@ var (
 	// GroupVersion is group version used to register these objects
 	GroupVersion = schema.GroupVersion{Group: "servicebinding.io", Version: "v1alpha3"}
 
-	GroupVersionResource = GroupVersion.WithResource("servicebindings")
-
-	GroupVersionKind = GroupVersion.WithKind("ServiceBinding")
-
-	WorkloadResourceMappingGroupVersionResource = GroupVersion.WithResource("clusterworkloadresourcemappings")
-
-	WorkloadResourceMappingGroupVersionKind = GroupVersion.WithKind("ClusterWorkloadResourceMapping")
-
 	// SchemeBuilder is used to add go types to the GroupVersionKind scheme
 	SchemeBuilder = &scheme.Builder{GroupVersion: GroupVersion}
