+++ apis/spec/v1alpha3/servicebinding_types.go
@@ -17,9 +17,7 @@ limitations under the License.
 package v1alpha3
 
 import (
-	"errors"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/runtime/schema"
 )
 
 // ServiceBindingWorkloadReference defines a subset of corev1.ObjectReference with extensions
@@ -124,37 +122,6 @@ type ServiceBindingList struct {
 
 func init() {
 	SchemeBuilder.Register(&ServiceBinding{}, &ServiceBindingList{})
-}
-
-func (ref *ServiceBindingServiceReference) GroupVersionResource() (*schema.GroupVersionResource, error) {
-	return nil, errors.New("Resource undefined")
-}
-
-func (ref *ServiceBindingServiceReference) GroupVersionKind() (*schema.GroupVersionKind, error) {
-	typeMeta := &metav1.TypeMeta{Kind: ref.Kind, APIVersion: ref.APIVersion}
-	gvk := typeMeta.GroupVersionKind()
-	return &gvk, nil
-}
-
-func (ref *ServiceBindingWorkloadReference) GroupVersionResource() (*schema.GroupVersionResource, error) {
-	return nil, errors.New("Resource undefined")
-}
-
-func (ref *ServiceBindingWorkloadReference) GroupVersionKind() (*schema.GroupVersionKind, error) {
-	typeMeta := &metav1.TypeMeta{Kind: ref.Kind, APIVersion: ref.APIVersion}
-	gvk := typeMeta.GroupVersionKind()
-	return &gvk, nil
-}
-
-func (sb *ServiceBinding) AsOwnerReference() metav1.OwnerReference {
-	var ownerRefController bool = true
-	return metav1.OwnerReference{
-		Name:       sb.Name,
-		UID:        sb.UID,
-		Kind:       sb.Kind,
-		APIVersion: sb.APIVersion,
-		Controller: &ownerRefController,
-	}
 }
 
 func (sb *ServiceBinding) HasDeletionTimestamp() bool {
