+++ apis/spec/v1beta1/groupversion_info.go
@@ -32,6 +32,10 @@ var (
 	// GroupVersion is group version used to register these objects
 	GroupVersion = schema.GroupVersion{Group: "servicebinding.io", Version: "v1beta1"}
 
+	GroupVersionResource = GroupVersion.WithResource("servicebindings")
+
+	GroupVersionKind = GroupVersion.WithKind("ServiceBinding")
+
 	// SchemeBuilder is used to add go types to the GroupVersionKind scheme
 	SchemeBuilder = &scheme.Builder{GroupVersion: GroupVersion}
