+++ test/_projects/api-client/main.go
@@ -3,7 +3,7 @@ package main
 import (
 	bindingapi "github.com/redhat-developer/service-binding-operator/apis/binding/v1alpha1"
 
-	specv1alpha2 "github.com/redhat-developer/service-binding-operator/apis/spec/v1alpha3"
+	specv1alpha2 "github.com/redhat-developer/service-binding-operator/apis/spec/v1beta1"
 	"k8s.io/apimachinery/pkg/runtime"
 )
