+++ test/_projects/api-controller/main.go
@@ -4,7 +4,7 @@ import (
 	"context"
 	bindingapi "github.com/redhat-developer/service-binding-operator/apis/binding/v1alpha1"
 
-	specv1alpha2 "github.com/redhat-developer/service-binding-operator/apis/spec/v1alpha3"
+	specv1alpha2 "github.com/redhat-developer/service-binding-operator/apis/spec/v1beta1"
 	"k8s.io/apimachinery/pkg/runtime"
 	utilruntime "k8s.io/apimachinery/pkg/util/runtime"
