+++ test/charts/service-binding-operator/sbo.yaml
@@ -1,4 +1,4 @@
-apiVersion: servicebinding.io/v1alpha3
+apiVersion: servicebinding.io/v1beta1
 kind: ServiceBinding
 metadata:
   name: test-sbo-chart-binding
