+++ test/performance/user-workloads/invalid/sbo-test.sbr.spec.yaml
@@ -1,5 +1,5 @@
 ---
-apiVersion: servicebinding.io/v1alpha3
+apiVersion: servicebinding.io/v1beta1
 kind: ServiceBinding
 metadata:
   name: service-binding-cron-invalid
