+++ test/performance/user-workloads/valid/sbo-test.with-sbr.user-workloads.yaml
@@ -62,7 +62,7 @@ objects:
         app: sbo-perf-svc-valid
     spec:
       replicas: 1
-      strategy: 
+      strategy:
         type: RollingUpdate
       selector:
         matchLabels:
@@ -162,7 +162,7 @@ objects:
           - watch
           - update
           - patch
-  - apiVersion: servicebinding.io/v1alpha3
+  - apiVersion: servicebinding.io/v1beta1
     kind: ClusterWorkloadResourceMapping
     metadata:
       name: cronjobs.batch
@@ -173,7 +173,7 @@ objects:
           containers:
             - path: .spec.jobTemplate.spec.template.spec.containers[*]
             - path: .spec.jobTemplate.spec.template.spec.initContainers[*]
-  - apiVersion: servicebinding.io/v1alpha3
+  - apiVersion: servicebinding.io/v1beta1
     kind: ServiceBinding
     metadata:
       name: service-binding-cron-valid
